Major changes to the theme
hardcoded the emails in the new partials for contact us and sign up here - TODO don't hard code this and just add this as a site param. Also made a few changes to the overall layout. Finally figured out how to get clicks on the logo to ACTUALLY go back to the home page.